Fixed Windows build
